refactor: move MainContent logic to a new component file for better separation of concerns
